Module empty::parser

Module parser is generated by GoGLL. Do not edit.

Modules

bsr

slot
Module slot is generated by gogll. Do not edit.

symbols

Structs

CRFNode

ClusterNode

Descriptor

Error
Error is returned by Parse at every point at which the parser fails to parse a grammar production. For non-LL-1 grammars there will be an error for each alternate attempted by the parser.

FIRST

FOLLOW

ParseError

Parser

PoppedNode

Enums

Expected

Functions

parse
Parse returns the BSR set containing the parse forest. If the parse was successfull the length of Vec<Box<Error>> = 0.