Struct empty::parser::bsr::Set

pub struct Set {
    slot_entries: HashMap<Rc<BSR>, bool>,
    nt_slot_entries: HashMap<NTSlot, Vec<Rc<BSR>>>,
    string_entries: HashMap<Rc<BSR>, bool>,
    pub rext: usize,
    lex: Rc<Lexer>,
    start_sym: NT,
}

Set contains the set of Binary Subtree Representations (BSR).

Fields
slot_entries: HashMap<Rc<BSR>, bool>nt_slot_entries: HashMap<NTSlot, Vec<Rc<BSR>>>string_entries: HashMap<Rc<BSR>, bool>rext: usizelex: Rc<Lexer>start_sym: NT
Methods

impl Set

pub fn new(start_symbol: NT, l: Rc<Lexer>) -> Box<Set>

New returns a new initialised BSR Set

pub fn add(&mut self, l: Label, i: usize, k: usize, j: usize)

Add a BSR to the set. (i,j) is the extent. k is the pivot.

pub fn alternate(&self, b: Rc<BSR>) -> usize

Returns the index of the grammar rule alternate.

fn insert(&mut self, bsr: Kind)

pub fn add_empty(&mut self, l: Label, i: usize)

AddEmpty adds a grammar slot: X : ϵ•

pub fn contain(&self, nt: &NT, left: usize, right: usize) -> bool

contain returns true iff the BSR Set contains the NT symbol with left and right extent.

pub fn get_all(&self) -> Vec<Rc<BSR>>

Returns all the NT BSR entries. Used for debugging.

pub fn get_root(&self) -> Rc<BSR>

pub fn get_roots(&self) -> Vec<Rc<BSR>>

fn get_line_column(&self, i: usize) -> (usize, usize)

Return the (line, column) of the left extent of token i.

pub fn get_nt_child_i(&self, b: Rc<BSR>, i: usize) -> Rc<BSR>

pub fn get_nt_children_i(&self, b: Rc<BSR>, i: usize) -> &Vec<Rc<BSR>>

fn get_nt_slot(&self, sym: &Symbol, lext: usize, rext: usize) -> &Vec<Rc<BSR>>

fn get_string(&self, l: Label, lext: usize, rext: usize) -> Rc<BSR>

pub fn get_t_child_i(&self, b: Rc<BSR>, i: usize) -> Rc<Token>

GetTChildI returns the terminal symbol at position i in b.
GetTChildI panics if symbol i is not a valid terminal

pub fn is_ambiguous(&self) -> bool

Returns true if the BSR set does not have exactly one root, or if any BSR in the set has an NT symbol, which does not have exactly one sub-tree.

fn is_ambiguous_bsr(&self, b: Rc<BSR>) -> bool

Returns true if b or any of its NT children is ambiguous. A BSR is ambigous if any of its NT symbols does not have exactly one subtrees (children).