Module empty::token

Module token is generated by GoGLL. Do not edit

Structs

STRING_TO_TYPE

TYPE_TO_STRING

Token
Token is returned by the lexer for every scanned lexical token

Enums

Type

Functions

new
New returns a new token.
lext is the left extent and rext the right extent of the token in the input.
input is the input slice scanned by the lexer.